Add manpage support
