+++ crates/bws/Cargo.toml
@@ -44,5 +44,11 @@ uuid = { version = "^1.7.0", features = ["serde"] }
 
 bitwarden = { workspace = true, features = ["secrets"] }
 
+[build-dependencies]
+clap = { version = "4.5.1", features = ["derive", "string"] }
+clap_complete = "4.5.0"
+clap_mangen = "0.2.20"
+uuid = { version = "^1.7.0" }
+
 [dev-dependencies]
 tempfile = "3.10.0"
