+++ crates/bws/src/config.rs
@@ -4,10 +4,11 @@ use std::{
     path::{Path, PathBuf},
 };
 
-use clap::ValueEnum;
 use color_eyre::eyre::{bail, Result};
 use directories::BaseDirs;
 use serde::{Deserialize, Serialize};
+
+use crate::cli::{ProfileKey, DEFAULT_CONFIG_DIRECTORY, DEFAULT_CONFIG_FILENAME};
 
 #[derive(Debug, Serialize, Deserialize, Default)]
 pub(crate) struct Config {
@@ -22,16 +23,6 @@ pub(crate) struct Profile {
     pub state_file_dir: Option<String>,
 }
 
-// TODO: This could probably be derived with a macro if we start adding more fields
-#[allow(non_camel_case_types)]
-#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
-pub(crate) enum ProfileKey {
-    server_base,
-    server_api,
-    server_identity,
-    state_file_dir,
-}
-
 impl ProfileKey {
     fn update_profile_value(&self, p: &mut Profile, value: String) {
         match self {
@@ -43,13 +34,13 @@ impl ProfileKey {
     }
 }
 
-pub(crate) const FILENAME: &str = "config";
-pub(crate) const DIRECTORY: &str = ".bws";
-
 pub(crate) fn get_config_path(config_file: Option<&Path>, ensure_folder_exists: bool) -> PathBuf {
     let config_file = config_file.map(ToOwned::to_owned).unwrap_or_else(|| {
         let base_dirs = BaseDirs::new().unwrap();
-        base_dirs.home_dir().join(DIRECTORY).join(FILENAME)
+        base_dirs
+            .home_dir()
+            .join(DEFAULT_CONFIG_DIRECTORY)
+            .join(DEFAULT_CONFIG_FILENAME)
     });
 
     if ensure_folder_exists {
