+++ crates/bws/src/render.rs
@@ -1,26 +1,9 @@
 use bitwarden::secrets_manager::{projects::ProjectResponse, secrets::SecretResponse};
 use chrono::{DateTime, Utc};
-use clap::ValueEnum;
 use comfy_table::Table;
 use serde::Serialize;
 
-#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
-#[allow(clippy::upper_case_acronyms)]
-pub(crate) enum Output {
-    JSON,
-    YAML,
-    Env,
-    Table,
-    TSV,
-    None,
-}
-
-#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
-pub(crate) enum Color {
-    No,
-    Yes,
-    Auto,
-}
+use crate::cli::{Color, Output};
 
 impl Color {
     pub(crate) fn is_enabled(self) -> bool {
